requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
coverage==5.2
django-anymail==8.1
django-cors-headers==3.5.0
django-money==1.1
django-versatileimagefield==2.0
djangorestframework==3.11.0
Django==3.1.3
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
py-moneyed==0.8.0
pylint==2.6.0
python-magic==0.4.18
psycopg2==2.8.6
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
stripe==2.51.0
toml==0.10.1
urllib3==1.25.10
wincertstore==0.2
wrapt==1.12.1
